+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to correctly prioritize next spottable elements when wrapped by a container that does not also wrap the currently focused element
+
 ## [3.4.4] - 2020-08-17
 
 No significant changes.